Fixed not checking for null.

+++ app/src/main/java/com/doctoror/fuckoffmusicplayer/playlist/PlaylistActivity.java
@@ -66,7 +66,7 @@ public final class PlaylistActivity extends BaseActivity implements
             mDeleteSession = state.deleteSession;
             mFinishWhenDialogDismissed = state.finishWhenDialogDismissed;
 
-            if (mDeleteSession.permissionRequested) {
+            if (mDeleteSession != null && mDeleteSession.permissionRequested) {
                 onDeleteClick(mDeleteSession.media);
             }
         }
